+++ spec/active_interaction/filters/hash_filter_spec.rb
@@ -51,10 +51,32 @@ describe ActiveInteraction::HashFilter, :filter do
           end
         end
       end
+
+      context 'without a Hash' do
+        let(:k) { 'a' }
+        let(:v) { double }
+        let(:value) { { k => v } }
+
+        it 'raises an error' do
+          expect do
+            filter.cast(value)
+          end.to raise_error ActiveInteraction::InvalidNestedValueError
+        end
+
+        it 'populates the error' do
+          begin
+            filter.cast(value)
+          rescue ActiveInteraction::InvalidNestedValueError => e
+            expect(e.filter_name).to eq k
+            expect(e.input_value).to eq v
+          end
+        end
+      end
     end
 
     context 'keys are symbolized' do
       let(:value) { { 'a' => 'a', 1 => 1 } }
+
       before do
         options.merge!(strip: false)
       end